Allow for a more fine grained error handling.
This is specially useful when using the same link worker for several requests (without HTTP pipelining).
In this particular use case, the client code wants to distinguish "real" errors from the socket closed (by the peer) error,
so that it either spawns a new link worker process and retries a request or it executes some error report/recovery code.